requirements.txt
requests
plyer
sv-ttk
pystray>=0.19.5
Pillow>=10.0.0
pyinstaller>=6.0.0